+++ src/sys.rs
@@ -1,9 +1,14 @@
-#![allow(non_upper_case_globals)]
-#![allow(non_camel_case_types)]
-#![allow(non_snake_case)]
-#![allow(dead_code)]
-#![allow(improper_ctypes)]
-#![allow(unnecessary_transmutes)]
-#![allow(clippy::all)]
+//! FDK AAC の FFI バインディング
+//!
+//! bindgen が生成したコードを include するためのモジュール。
+//! 生成コードはリポジトリ管理外のため、lint の抑制には
+//! `#[expect(...)]` を使い、警告が発生しなくなったら
+//! 抑制自体を削除する必要があることに気づけるようにする。
+
+#![expect(non_upper_case_globals)]
+#![expect(non_camel_case_types)]
+#![expect(non_snake_case)]
+#![expect(dead_code)]
+#![expect(clippy::all)]
 
 include!(concat!(env!("OUT_DIR"), "/bindings.rs"));